artha/services/artha-api/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
uvloop==0.22.1
httptools==0.7.1
asyncpg==0.31.0
pydantic==2.12.5
pydantic-settings==2.13.1
orjson==3.11.7
python-jose[cryptography]==3.5.0
httpx==0.28.1
pandas==3.0.1
numpy==2.4.3
psycopg[binary]==3.3.3
slowapi==0.1.9
cachetools==7.0.5
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-mock==3.15.1
